refactor: fix name files, func
