+++ add-lot.php
@@ -6,8 +6,8 @@ require_once ('init.php');
 require_once ('utils/helpers.php');
 require_once ('models/categories.php');
 require_once ('models/lots.php');
-require_once('validate/validate.php');
-require_once('validate/validate_upload_file.php');
+require_once('validate/add-lot.php');
+require_once('validate/upload-file.php');
 
 /**
  * @var boolean|mysqli|object $connect ресурс соединения с сервером БД
