+++ templates/layout.php
@@ -32,7 +32,7 @@ declare(strict_types=1);
                 <input type="search" name="search" placeholder="Поиск лота">
                 <input class="main-header__search-btn" type="submit" name="find" value="Найти">
             </form>
-            <a class="main-header__add-lot button" href="<?=create_new_url('add.php'); ?>">Добавить лот</a>
+            <a class="main-header__add-lot button" href="<?=create_new_url('add-lot.php'); ?>">Добавить лот</a>
 
             <nav class="user-menu">
