+++ utils/date-time.php
@@ -6,48 +6,24 @@ const SECS_IN_HOUR = 3600;
 // Количество секунд в одной минуте
 const SECS_IN_MINUTE = 60;
 
-
-
 /**
-
  * Проверяет переданную дату на соответствие формату 'ГГГГ-ММ-ДД'
-
  *
-
  * Примеры использования:
-
  * is_date_valid('2019-01-01'); // true
-
  * is_date_valid('2016-02-29'); // true
-
  * is_date_valid('2019-04-31'); // false
-
  * is_date_valid('10.10.2010'); // false
-
  * is_date_valid('10/10/2010'); // false
-
  *
-
  * @param string $date Дата в виде строки
-
- *
-
  * @return bool true при совпадении с форматом 'ГГГГ-ММ-ДД', иначе false
-
  */
-
 function is_date_valid(string $date) : bool {
-
     $format_to_check = 'Y-m-d';
-
     $dateTimeObj = date_create_from_format($format_to_check, $date);
-
-
-
     return $dateTimeObj !== false && array_sum([date_get_last_errors()]) === 0;
-
 }
-
 
 /**
  * Функция возвращает остаток времени до данной даты в виде массива
@@ -62,15 +38,11 @@ function time_diff(string $date) : array
     $ts_end_time = strtotime($date);
     $ts_diff = $ts_end_time - $ts;
     if ($ts_diff < 0) {
-
         $hours = 0;
         $minutes = 0;
-
     } else {
-
         $hours = floor($ts_diff/ SECS_IN_HOUR);
         $minutes = floor(($ts_diff % SECS_IN_HOUR) / SECS_IN_MINUTE);
-
     }
     return [
         'hours' => $hours,
@@ -83,7 +55,6 @@ function time_diff(string $date) : array
  *
  * @param string $date дата в формате «ГГГГ-ММ-ДД»
  * @return float
- *
  **/
 function diff_date(string $date) : float
 {
